feat: add mappedArgs configuration to setupTestWiring
When confirm the right arguments are passed into an api call, a frequent
problem is that you only real care about some small subset in the
tests(passing in the whole redux store, etc).  If you're only confirming
those arguments, its easy enough to clean them up in the test, but if
you have to confirm multiple times, mappedArgs is a much better
option.  By setting the key in mappedArgs to the function who's args
you want to update, you can perform the clean up once and have it be
shared across all waitFor calls