refactor: storybook render in tests
There's no real way to test the code the wrapping of
decorators that happens when storybook runs.  Restructuring
the add decorator code was closest we could get.  By manually
wrapping the same function that is evetually going to be called
by addDecorator in preview.js, we can test everything but the
actual creation of the decorator.

Achieving that required moving some functions around to make it
possible to get access to the same wrapper in tests