feat: added the ability to mock handlers
A frequent task when testing components is confirming that
passed in handler props (onClick, onClose, etc) were called correctly
at the right times with the right arguments. This adds the ability
to pass a handlers array to setupTestWiring that will
automatically make it possible to wait for all of the prop names
that correspond to names in handlers.

This also adds a createStory convenience helper
to create stories